Refactor CLI TUI package layout
