+++ kolega_code/cli/tui/__init__.py
@@ -1,0 +1,1 @@
+"""Textual UI internals for the Kolega Code CLI."""
