+++ kolega_code/cli/tui/constants.py
@@ -1,0 +1,24 @@
+"""Constants used by the Textual UI."""
+
+BUILD_INTERACTION_MODE = "build"
+PLAN_INTERACTION_MODE = "plan"
+QUESTION_TOOL_NAME = "ask_user_choice"
+QUESTION_OPTION_ID_PREFIX = "question_option_"
+APPROVAL_OPTION_ID_PREFIX = "approval_option_"
+MODEL_OPTION_ID_PREFIX = "model_option_"
+EFFORT_OPTION_ID_PREFIX = "effort_option_"
+THEME_OPTION_ID_PREFIX = "theme_option_"
+
+TAB_BASE_LABELS = {
+    "logs_pane": "Logs",
+    "terminal_pane": "Terminal",
+}
+
+STARTUP_WORDMARK = (
+    " _  __     _                   ____          _",
+    "| |/ /___ | | ___  __ _  __ _ / ___|___   __| | ___",
+    "| ' // _ \\| |/ _ \\/ _` |/ _` | |   / _ \\ / _` |/ _ \\",
+    "| . \\ (_) | |  __/ (_| | (_| | |__| (_) | (_| |  __/",
+    "|_|\\_\\___/|_|\\___|\\__, |\\__,_|\\____\\___/ \\__,_|\\___|",
+    "                  |___/",
+)
